extract kernel package
