+++ .gitignore
@@ -5,6 +5,10 @@ tmp
 
 database/infra/data
 
+# --- [API]: Bin
+bin/*.*
+!bin/.gitkeep
+
 # --- [API]: Storage
 storage/logs/*.*
 storage/media/*.*
