+++ database/seeder/main.go
@@ -1,20 +1,21 @@
 package main
 
 import (
-	"github.com/oullin/metal"
+	"sync"
+	"time"
+
 	"github.com/oullin/database"
 	"github.com/oullin/database/seeder/seeds"
 	"github.com/oullin/env"
+	"github.com/oullin/metal/kernel"
 	"github.com/oullin/pkg"
 	"github.com/oullin/pkg/cli"
-	"sync"
-	"time"
 )
 
 var environment *env.Environment
 
 func init() {
-	secrets := metal.Ignite("./.env", pkg.GetDefaultValidator())
+	secrets := kernel.Ignite("./.env", pkg.GetDefaultValidator())
 
 	environment = secrets
 }
@@ -22,8 +23,8 @@ func init() {
 func main() {
 	cli.ClearScreen()
 
-	dbConnection := metal.MakeDbConnection(environment)
-	logs := metal.MakeLogs(environment)
+	dbConnection := kernel.MakeDbConnection(environment)
+	logs := kernel.MakeLogs(environment)
 
 	defer (*logs).Close()
 	defer (*dbConnection).Close()
