+++ metal/kernel/app.go
@@ -1,7 +1,9 @@
-package metal
+package kernel
 
 import (
 	"fmt"
+	baseHttp "net/http"
+
 	"github.com/oullin/database"
 	"github.com/oullin/database/repository"
 	"github.com/oullin/env"
@@ -9,7 +11,6 @@ import (
 	"github.com/oullin/pkg/auth"
 	"github.com/oullin/pkg/http/middleware"
 	"github.com/oullin/pkg/llogs"
-	baseHttp "net/http"
 )
 
 type App struct {
